requirements.txt
aiogram==3.13.1
pyrogram==2.0.106
tgcrypto==1.2.5
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.0
cryptography==44.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-generativeai==0.8.3
uvloop==0.21.0
